Add scaffolding and initial download process
