+++ etl-pipeline/scripts/__init__.py
@@ -18,3 +18,4 @@ from .deleteProblemWorks import main as deleteWorks
 from .report_missing_files import main as ReportMissingFiles
 from .create_collection import main as createCollection
 from .GRIN_initial_scrape import main as GRINInitialScrape
+from .grin_file_acquisition import main as GRINFileAcquisition
